refactor: Pagination 컴포넌트 통합 (fraction/list → type prop으로 일원화)
