+++ src/components/Pagination/Pagination.tsx
@@ -1,23 +1,32 @@
 import { ArrowLeftIcon } from "@/components/Icon/ArrowLeftIcon";
 import { ArrowRightIcon } from "@/components/Icon/ArrowRightIcon";
+import { ChevronLeftIcon } from "@/components/Icon/ChevronLeftIcon";
+import { ChevronRightIcon } from "@/components/Icon/ChevronRightIcon";
 import { usePagination } from "@/hooks/usePagination";
 import { cn } from "@/lib/utils/utils";
 
 import { ELLIPSIS, getPageSlots } from "./Pagination.utils";
 
-interface PaginationListProps {
+interface PaginationBaseProps {
   totalPage: number;
   currentPage: number;
   onPageChange: (page: number) => void;
   className?: string;
 }
 
-export function PaginationList({
-  totalPage,
-  currentPage,
-  onPageChange,
-  className,
-}: PaginationListProps) {
+interface PaginationFractionProps extends PaginationBaseProps {
+  type: "fraction";
+}
+
+interface PaginationListProps extends PaginationBaseProps {
+  type: "list";
+}
+
+type PaginationProps = PaginationFractionProps | PaginationListProps;
+
+export function Pagination(props: PaginationProps) {
+  const { totalPage, currentPage, onPageChange, className, type } = props;
+
   const {
     normalizedTotalPage,
     normalizedCurrentPage,
@@ -36,6 +45,45 @@ export function PaginationList({
     return null;
   }
 
+  if (type === "fraction") {
+    return (
+      <div className={cn("gap-sm flex items-center justify-center", className)}>
+        <button
+          type="button"
+          onClick={goToPreviousPage}
+          disabled={isFirstPage}
+          className={cn(
+            "flex size-6 shrink-0 items-center justify-center rounded-full p-1 transition-colors md:size-10",
+            isFirstPage ? "cursor-not-allowed opacity-50" : "hover:bg-surface-strong cursor-pointer"
+          )}
+          aria-label="Previous page"
+        >
+          <ChevronLeftIcon className="text-text-disabled size-4 md:size-6" />
+        </button>
+
+        <div className="gap-2xs text-text-muted flex items-center justify-center text-base font-semibold">
+          <span>{normalizedCurrentPage}</span>
+          <span>/</span>
+          <span>{normalizedTotalPage}</span>
+        </div>
+
+        <button
+          type="button"
+          onClick={goToNextPage}
+          disabled={isLastPage}
+          className={cn(
+            "flex size-6 shrink-0 items-center justify-center rounded-full p-1 transition-colors md:size-10",
+            isLastPage ? "cursor-not-allowed opacity-50" : "hover:bg-surface-strong cursor-pointer"
+          )}
+          aria-label="Next page"
+        >
+          <ChevronRightIcon className="text-text-disabled size-4 md:size-6" />
+        </button>
+      </div>
+    );
+  }
+
+  // type === "list"
   const pages = getPageSlots(normalizedTotalPage, normalizedCurrentPage);
 
   return (
